moved modal to the admin page
