+++ src/pages/PlayerPage/PlayerPage.js
@@ -2,17 +2,9 @@ import React, { useState, useEffect } from 'react';
 import axios from 'axios';
 import './PlayerPage.scss';
 import PlayerListItem from '../common/PlayerListItem/PlayerListItem.js';
-import Button from '../common/Button/Button';
-import AddPlayerForm from '../AddPlayerForm/AddPlayerForm';
 
 const PlayerPage = () => {
     const [players, setPlayers] = useState([]);
-    const [addPlayerModal, setAddPlayerModal] = useState(false);
-    const [newPlayerValues, setNewPlayerValues] = useState({
-        name: "",
-        team: "",
-        position: ""
-    });
 
     useEffect(() => {
         const fetchData = async () => {
@@ -22,19 +14,9 @@ const PlayerPage = () => {
         fetchData();
     }, []);
 
-    const submitForm = () => {
-        const { data } = axios.post('/player/')
-        setAddPlayerModal(false)
-    }
-
-    const onFormChange = (value, inputName) => {
-        setNewPlayerValues({...newPlayerValues, [inputName]: value})
-    }
-
     return (
         <div className="Playerpage">
             <div className="Playerpage-body">
-                <Button onClick={() => setAddPlayerModal(true)}>Add Players</Button>
                 <div className="player-list">
                 {
                     players.map((player, key) => {
@@ -44,9 +26,6 @@ const PlayerPage = () => {
                         })
                     }
                 </div>
-                {
-                    addPlayerModal && (<AddPlayerForm closeModalFn={setAddPlayerModal} changeFn={onFormChange} values={newPlayerValues} submitFn={submitForm}/>) 
-                }
             </div>
         </div>
     );
